+++ CAESAR/src/scheduler/RunAndQueryDrivenScheduler.java
@@ -24,7 +24,7 @@ public class RunAndQueryDrivenScheduler extends Scheduler implements Runnable {
 	int HP_frequency;
 	int LP_frequency;
 	
-	public RunAndQueryDrivenScheduler (AtomicInteger dp, HashMap<Double,Double> distrProgrPerSec, HashMap<RunID,Run> rs, RunQueues rq, RunQueues hprq,ExecutorService e, 
+	public RunAndQueryDrivenScheduler (AtomicInteger dp, HashMap<Double,Long> distrProgrPerSec, HashMap<RunID,Run> rs, RunQueues rq, RunQueues hprq,ExecutorService e, 
 								CountDownLatch tn, CountDownLatch d, ArrayList<XwayDirPair> xds, int lastS, long start, AtomicInteger x0, AtomicInteger x1, int hpf, int lpf) {		
 		super(dp,distrProgrPerSec,rs,rq,e,tn,d,xds,lastS,start);
 		HPrunqueues = hprq;
